Moved Board type from fullgame -> boardinit
