+++ src/client/scripts/esm/game/boardeditor/tools/normaltool.ts
@@ -8,8 +8,9 @@
 
 import type { Mesh } from '../../rendering/piecemodels';
 import type { Edit } from '../../../../../../shared/chess/logic/movepiece';
+import type { Board } from '../../../../../../shared/chess/logic/boardinit';
+import type { FullGame } from '../../../../../../shared/chess/logic/fullgame';
 import type { MoveCoords } from '../../../../../../shared/chess/logic/icn/icnconverter';
-import type { Board, FullGame } from '../../../../../../shared/chess/logic/fullgame';
 
 import state from '../../../../../../shared/chess/logic/state';
 import movepiece from '../../../../../../shared/chess/logic/movepiece';
