+++ src/client/scripts/esm/game/rendering/highlights/legalmovemodel.ts
@@ -8,10 +8,11 @@
  */
 
 import type { Color } from '../../../../../../shared/util/math/math.js';
+import type { Board } from '../../../../../../shared/chess/logic/boardinit.js';
 import type { Player } from '../../../../../../shared/chess/util/typeutil.js';
+import type { FullGame } from '../../../../../../shared/chess/logic/fullgame.js';
 import type { MoveTagged } from '../../../../../../shared/chess/logic/movepiece.js';
 import type { IgnoreFunction } from '../../../../../../shared/chess/logic/movesets.js';
-import type { Board, FullGame } from '../../../../../../shared/chess/logic/fullgame.js';
 import type { Ray, Vec2, Vec2Key } from '../../../../../../shared/util/math/vectors.js';
 import type { LegalMoves, SlideLimits } from '../../../../../../shared/chess/logic/legalmoves.js';
 import type {
