+++ src/client/scripts/esm/game/rendering/pieces.ts
@@ -6,7 +6,7 @@
  */
 
 import type { Mesh } from './piecemodels.js';
-import type { Board } from '../../../../../shared/chess/logic/fullgame.js';
+import type { Board } from '../../../../../shared/chess/logic/boardinit.js';
 import type { Coords } from '../../../../../shared/chess/util/coordutil.js';
 
 import meshes from './meshes.js';
