+++ src/shared/chess/logic/fourdimensionalmoves.ts
@@ -9,13 +9,14 @@
  */
 
 import type { Piece } from '../util/boardutil.js';
+import type { Board } from './boardinit.js';
 import type { Coords } from '../util/coordutil.js';
 import type { Player } from '../util/typeutil.js';
 import type { Dimensions } from '../variants/variant_scripts/gen4DPosition.js';
 import type { MoveRunning } from './specialmove.js';
 import type { CoordsTagged } from './movepiece.js';
+import type { Game, FullGame } from './fullgame.js';
 import type { UnboundedRectangle } from '../../util/math/bounds.js';
-import type { Game, Board, FullGame } from './fullgame.js';
 
 import state from './state.js';
 import bimath from '../../util/math/bimath.js';
