+++ src/shared/chess/logic/movepiece.ts
@@ -7,13 +7,14 @@
  */
 
 import type { Piece } from '../util/boardutil.js';
+import type { Board } from './boardinit.js';
 import type { Coords } from '../util/coordutil.js';
 import type { Change } from './boardchanges.js';
+import type { FullGame } from './fullgame.js';
 import type { MoveState } from './state.js';
 import type { MoveCoords } from './icn/icnconverter.js';
 import type { MovePacket } from '../../types.js';
 import type { GameConclusion } from '../util/winconutil.js';
-import type { Board, FullGame } from './fullgame.js';
 import type { MoveSpecialTags, SpecialTags } from '../util/moveutil.js';
 
 import state from './state.js';
